Elastic Search Logging with Open API additions

+++ Plutus.Application/Transactions/Commands/Delete.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using FluentValidation;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using Plutus.Application.Exceptions;
 using Plutus.Application.Repositories;
 using Plutus.Application.Transactions.Indexes;
 using Plutus.Domain.ValueObjects;
@@ -12,12 +13,12 @@ namespace Plutus.Application.Transactions.Commands
 {
     public static class DeleteTransaction
     {
-        public record Request(Guid Id) : IRequest<Response>;
+        public record Request(Guid Id) : IRequest<(AbstractPlutusException? error, Response? response)>;
 
         public record Response(Guid Id);
 
 
-        public class Handler : IRequestHandler<Request, Response>
+        public class Handler : IRequestHandler<Request, (AbstractPlutusException? error, Response? response)>
         {
             private readonly ITransactionRepository _repository;
             private readonly TransactionIndex _transactionIndex;
@@ -28,15 +29,18 @@ namespace Plutus.Application.Transactions.Commands
                 _transactionIndex = transactionIndex;
             }
 
-            public async Task<Response> Handle(Request request, CancellationToken cancellationToken)
+            public async Task<(AbstractPlutusException? error, Response? response)> Handle(Request request, CancellationToken cancellationToken)
             {
                 var transactionToDelete = await _repository.FindByIdAsync(request.Id);
+
+                if (transactionToDelete.InActive || transactionToDelete is null) return (new TransactionNotFoundException(), null);
+
                 transactionToDelete.MakeInActive();
                 _repository.Update(transactionToDelete);
                 await _repository.SaveChangesAsync();
                 
                 await _transactionIndex.DeleteAsync(request.Id.ToString(), cancellationToken);
-                return new Response(request.Id);
+                return (null, new Response(request.Id));
             }
         }
